Refix Travis references to branches and tags.
Referring to origin/* is wrong, because that doesn't work for tags.
In fact, there is no way we can refer to tags other than a bare
string, and since we need to use branches and tags interchangably, we
have to make branches follow the same pattern.

Therefore, we have to do the Git magic in this patch to fetch the
branches into local branches with bare names.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>