Update license check to check for specific copyright year.
License test is checking now when the source file was modified and
makes sure the last modification date matches the copyright information
inside the source file.
This way all the modified files need to have the license updated, but
the ones which are not modified can stay with the previous copyright
information.

Changelog: None

Signed-off-by: Marcin Pasinski <marcin.pasinski@northern.tech>