MEN-2674: Add support for Enterprise licenses.
Repositories that use Enterprise licenses are expected to use
`SetFirstEnterpriseCommit()` to set the first commit SHA that is
Enterprise-only. This will in turn be passed using the
`--ent-start-commit` flag to the `check_license_go_code.sh` script,
which is used to calculate which files are in Enterprise-only
(Enterprise license), or only in Open Source or both (Open Source
license).

Also took the chance to clean up the output of the script a little
bit, so that it no longer reports passing files. This makes it much
easier to see which files are affected.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>