QA-111: Check signed commits on community repositories
This is done through adding two flags to the chech_commits.sh script.
One for enabling the signature check, and one for enabling the Changelog check.

In the case that no flags are given, they will both be checked (The old
behaviour).

Also two files are added to be included by the respective .gitlab-ci.yml files,
depending on whether they require only the changelog, or the signoff checks.

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>