QA-131: Add a .gitlab-ci file which can be included in other files
This commit adds a '.gitlab-ci-check-python3-format.yml' file which can be included
into other projects' .gitlab-ci file, and which, when included, will check that
the Python3 code in the given directories are formatted according to Black's
preferences.

The code is checked with the 'check_python_code_format' script, which runs
through all the commits in a PR and runs Black on all the Python code. A filter
can be added, which only runs the code on python files matching a directory
path. See the file 'check_python_code_format' for more information.

Changelog: Title

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>