Skip master execution of mendertesting commit checks
This is mainly for mender-qa, where there vast majority of times we are
using "master" and wasting precious GitLab CI minutes with commit
checks. It won't hurt saving minutes on other repos neither.

Strictly speaking, we could apply the same principle to all templates,
but let's limit it only to the commit ones: as these are the really
superfluous ones in master while the others could possibly detect
errors in ninja merges.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>