Rename acceptance tests tempalte to avoid "golang" keyword
Now that we have removed the golang-specific test_prep_build stage, this
template will really build/run any acceptance test framework, as long as
the tests and the software under tests have Dockerfile(s) to instruct
the CI what to do.

Rename it now while it is not (yet) widely in use.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>