Bugfix in COMMIT_TAG definition in docker-build template
The bug has been present since the conception of the template. However
few repos have been using it until now.

Without quoting or brackets, the shell was expanding COMMIT_TAG to only
CI_COMMIT_SHA (as CI_COMMIT_REF_SLUG_ would not be found). This was
making the published master image to miss the 'master_' prefix and the
master tag to not be updated accordingly.

As we don't use these images for production nor for testing, no problems
have been reported.

Updated also SERVICE_IMAGE for consistency, but it is not really
necessary there.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>