Extend acceptance template for enterprise tests and normalize file names
Most of the repos use docker-compose-acceptance[-enterprise].yml, so set
this as the default and non compliant (inventory and deviceauth). The
enterprise tests are optional, though.

For some repos (namely closed source ones) it is required to log in
registry.mender.io in order to run acceptance tests

Also, some repos (only tenantadm?) need to copy a key into the test
runner.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>