QA-161: Fail on MenderCompliance issues
Previously the CheckMenderCompliance function would return an error, instead of
failing. This did not align with the previous functionality, in which the
function would cause the test to fail.

This has now been reverted.

Also, the commit changes the running of the shell scripts used for verifying
licenses and commits to be run directly in bash, instead of as a script. This
change is required due to go mod's vendoring functionality, in which it will
remove the executable bits on all files in the vendor directory.

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>