[docker-build-template] REF_NAME instead of REF_SLUG for publish tags
So that branch 1.2.x of a given repo gets published as <repo>:1.2.x (and
mender-1.2.x) instead of 1-2-x.

Renamed a bit the env variables to clarify their usage.

Obs! This will not work for publishing branches with slash (like
feature/my-coool-stuff) but that we can fix later on.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>