fix: Make the 'check_license.sh' script fail on wrong shasum lines
Previously, a misformed line, like a line with a sha that was missing a
character would pass the check, as the program would only produce a warning, and
not an error code.

This fix makes the program fail on such formatting errors.

Note that it does the verification on a temp-file, as the input license file
needs to be stripped of newlines, as these are also reported as formatting errors.

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>