Fix Enterprise license checker when OS has never been merged in.
It's a bit of a corner case. For the purposes of this explanation, let
the first Enterprise commit be E, and the current latest Enterprise be
H. The expression:

```
git rev-list E..H --ancestry-path --boundary --date-order
```

lists every commit between E and H, except E itself. Open Source
commits are not included because `--ancestry-path` forces every commit
to be a direct descendant of E. We then use `--date-order` and
`--boundary` to also output boundary commits that are not in the
range, in date order. The topmost of these is therefore the latest
commit which is not in the E..H range.

But wait, we just said that the commit E is *not* part of the range,
and indeed this commit is output as a boundary commit instead of an
in-range commit. If there has been no new Open Source merges after
this, then E will float to the top of the the list of boundary
commits, and will fool the checker into thinking it is an Open Source
commit.

It turns out that Git is not able to both list E and use
`--ancestry-path` at the same time, so we have to filter it out from
the list of boundary commits using grep instead.

The reason this has worked for other repositories is that as soon as
the repository has at least one additional merge from Open Source, the
problem disappears, because there is now a boundary commit with a
later date.

Finally, since we are erasing one boundary commit, there is a slight
possibility that there are no boundary commits left, so take that also
into account (see comments in the code).

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>