QA-363: Add a commit linter capable of parsing conventional commits
This is a rather big commit, consisting of all the parts needed in order to
enforce our decided conventional commit specification.

The given specification can be found in the `commitlint/grammar.md` file, and
should be pretty comprehensive.

The actual linter is an AWK script, found in `commitlint/commitlint`, and
enforces the grammar described in the previously mentioned document.

This enforces our conventional commit specification everywhere the
`commit_check` currently runs (so all our repositories).

Unversioned repositories will have to add the `UNVERSIONED_REPOSITORY: true`
variable to their `.gitlab-ci.yml` files, in order to opt out of the check.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>