feat: call the rerun_build endpoint to finalize the coverage report
Coveralls.io implemented a new end-point (Rerun Build Webhook) to
trigger a rebuild of the coverage report when multiple builds are in
use.  Without a call to this end-point, the aggregated coverage report
for builds could miss data from one of the builds, leading to an
erroneous drop in coverage which can cause a failed check on GitHub.

Changelog: None
Ticket: None

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>