fix: Make commit tags case insensitive.
There is no reason for them to be case sensitive, and it's confusing
for contributors. Also, several old commits already carry tags like
these.

Note the added line-beginning anchor in the regex. This was always
wrong because it would match "Cancel-changelog" as well, but because
the test used a lowercase "c", it wasn't caught until we switched it
to be case insensitive.

Changelog: none
Ticket: none

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>