ci: Allow to fail Mender specific jobs on build/publish template
This template is used by many repositories that build and publish Docker
images. For Mender specific images, there are extra jobs that deal with
Mender specific use cases.

Ideally these jobs should be disabled (removed from the pipeline) but so
far we were only early exiting leaving the job "green".

With this change we return an specific error code to fail the job while
we allow for the whole pipeline to continue.

The motivation is to better visualize what the job has done (nothing)
instead of visualizing it as just success (while doing nothing).

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>