fix(commitlint): Fix the definitions of WORD and COMMIT_SCOPE
A word is not a sequence of arbitrary **printable**
characters. It's shouldn't contain dashes either, but in our case
it simplifies things. For example, commit scope can then be a
word in parantheses.

This fixes issues with commit messages containing scope and
parantheses in the title, e.g. when mentioning SomeFunction().

Ticket: None
Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>
Signed-off-by: Vratislav Podzimek <v.podzimek@mykolab.com>