fix: avoid docker login hangs when missing creds
Ticket: QA-752

Recently, a docker login failure is waiting for 15 min to wait for web
browser login. This slows down a lot the pipelines that don't need login
to docker, but they are reusing the mendertesting shared pipeline
configurations.
With this change, we'll ignore the docker login if the variable is not
set.

Signed-off-by: Roberto Giovanardi <roberto.giovanardi@northern.tech>