fix: Do not override EXTRA_DOCKER_ARGS from parent scope
Depending on the trigger of the build, the variable value from the
template will get precedence, clearing the intended value of the user.

This is the case for example when running a manual pipeline through GUI,
where this variable is blanked out. See a failure pipeline here:
* https://gitlab.com/Northern.tech/Mender/mender-convert/-/jobs/9696025686

Ticket: None
Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>