chore: Split templates into real CI components and suport functions
The real CI component is expected to follow certain criterias like for
example be self contained. Some of the jobs in the template directory
does not fit those requirements and serve the prupose of a common
functions instead.

Those jobs have been moved to a separated `qa-common` directory and
updated to fit template requirements.

Ticket: QA-1344
Signed-off-by: Paweł Poławski <pawel.polawski@northern.tech>