ci: Add additional CI job to `commit-lint` CI/CD component for signoffs
There is a know issue with `commitlint` tool in that does not verify
signoffs tags the way we need it. We have a task to follow-up and the
temporary solution is to use both in the different projects. See for
example:
* https://github.com/mendersoftware/mender-cli/commit/69dab678b2fe0676be0ed49b986dc12f651dadd7

The intention of adding the job here is to centralize the responsibility
for the signoff checks (it is a very easy thing to oversee when creating
a new repo) and to have a single place where to remove it later once we
have QA-1046 in place.

Ticket: QA-1046

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>